requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
black==23.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.38.0
coverage==7.2.7
Django==4.2.3
django-cleanup==8.1.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.2.0
django-filter==23.3
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
exceptiongroup==1.2.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
pluggy==1.2.0
psycopg2-binary==2.9.6
PyJWT==2.8.0
pytest==7.4.0
pytest-django==4.5.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.0
